pyproject.toml
[project]
name = "ctx-to-lora"
version = "0.0.1"
authors = [{ name = "Rujikorn Charakorn" }]
description = ""
readme = "README.md"
requires-python = ">= 3.10"
dependencies = [
    # Latest 4.x; v5 migration is a follow-up commit.
    "transformers>=4.57.6,<5.0.0",
    "accelerate>=1.13.0",
    "datasets>=3.6.0",
    "setuptools",
    "peft>=0.19.1",
    "jupyter",
    "matplotlib",
    "hf_transfer",
    "torchmetrics",
    "inflect",
    "pre-commit",
    "tensorboardX",
    "wandb",
    "einops",
    "jaxtyping",
    "tensorboard",
    "flask",
    "gradio>=4.40.0",
    "pandas",
    "plotly",
    "rouge-score",
    "huggingface-hub[hf-transfer]>=0.32.0",
    "opt-einsum>=3.4.0",
    "kagglehub[hf-datasets]>=0.3.12",
    "kaggle>=1.7.4.5",
    "bitsandbytes>=0.49.2",
    "google-cloud-storage>=3.2.0",
    "wonderwords>=2.2.0",
    "llmlingua>=0.2.2",
]

# Heavy / training-only optional extras. Kept out of the default install so
# Blackwell + cu13 users (which lacks prebuilt wheels for some of these) can
# install the inference path without compiling everything from source.
[project.optional-dependencies]
training = [
    "deepspeed>=0.17.1",
    "liger-kernel",
]
vllm = ["vllm>=0.19.1"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pyright]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.*",
    ".venv",
    ".github",
    ".vscode",
    "chat_templates",
    "eval_results",
    "configs",
    "EditingLlama",
    "icae_v2",
    "lm-evaluation-harness",
    "llm-comparator",
    "LongBench",
    "scripts",
    "train_outputs",
    "./data/",
    "/data/",
    "generated_tasks",
    "outputs",
    "plots",
    "tmp",
    "wandb",
    ".wandb",
    ".ruff_cache",
    "assets",
]
typeCheckingMode = "off"

[tool.ruff]
line-length = 88
select = ["F401"]   # remove unused imports
ignore = ["E", "F"]

[tool.isort]
profile = "black"
known_local_folder = ["ctx_to_lora"]
